Remove adapters modules, just expose them in core modules
