+++ src/core/backup/mod.rs
@@ -1,0 +1,5 @@
+mod backup_execution;
+mod backup_index;
+
+pub use backup_execution::BackupExecution;
+pub use backup_index::BackupIndex;
